renamed checkout-page to checkout and put all files in one folder
