--- src/app/checkout-page/page.tsx
@@ -1,5 +1,0 @@
-import Checkout from "./components/Checkout"
-
-export default function CheckoutPage() {
-    return (<Checkout />)
-}
