+++ src/app/checkout/EquipmentSelector.tsx
