+++ src/app/checkout/page.tsx
@@ -1,0 +1,5 @@
+import CheckoutPage from "./CheckoutPage"
+
+export default function CheckoutPageWrapper() {
+    return <CheckoutPage />;
+}
