+++ src/app/page.tsx
@@ -1,5 +1,3 @@
-import CheckoutPage from "./checkout-page/page";
-
 export default function Home() {
-  return <CheckoutPage />;
+  return <></>;
 }
